worked w ryan to add more changes to clean up reducers and its test
